refactor: simplify javascript architecture
Major changes include:

- Migrate postcss to plain css (and remove the compile step).
- Migrate custom code from using browserify + common-js to
  ES modules (and remove the bundle step).
- Use Snowpack to install ES-module-compatible 3rd party
  dependencies.
- Update custom third-party dependencies (PrismJS & Modernizr).
- Update linting tooling.
- Update contribution docs.

Fixes #77
