--- .babelrc
@@ -1,8 +1,0 @@
-{
-  "presets": [
-    "@babel/env"
-  ],
-  "plugins": [
-    "@babel/plugin-proposal-object-rest-spread"
-  ]
-}
