+++ .stylelintrc
@@ -1,3 +1,3 @@
 {
-  "extends": "stylelint-config-sparkbox"
+  "extends": "@sparkbox/stylelint-config-sparkbox"
 }
