+++ index.html
@@ -12,10 +12,10 @@
   <meta property="og:description" content="Bouncy Ball compares web animation techniques by showing how to bounce a ball with each one." />
   <meta property="og:type" content="website" />
   <meta property="og:url" content="https://sparkbox.github.io/bouncy-ball" />
-  <meta property="og:image" content="https://sparkbox.github.io/bouncy-ball/site/assets/images/graphic.png" />
+  <meta property="og:image" content="https://sparkbox.github.io/bouncy-ball/site/images/graphic.png" />
 
   <link href="https://fonts.googleapis.com/css?family=Source+Sans+Pro:400|Source+Code+Pro|Ubuntu:500" rel="stylesheet">
-  <link rel="stylesheet" href="site/assets/styles.css">
+  <link rel="stylesheet" href="site/css/styles.css">
 </head>
 
 <body class="no-focus">
@@ -64,6 +64,7 @@
       <iframe class="demo-frame" src="examples/vanilla-js/index.html" name="bouncy ball demo" title="bouncy ball demo" frameborder="1" scrolling="no"></iframe>
     </div>
     <div class="pane source-pane">
+      <div class="prism-show-language"></div>
       <pre><code class="language-javascript"></code></pre>
     </div>
   </div>
@@ -71,7 +72,7 @@
     <span>Made by your friends at
       <a class="credit-link" href="https://twitter.com/hearsparkbox">@hearsparkbox</a>.</span>
   </div>
-  <script src="site/assets/bundle.js"></script>
+  <script type="module" src="site/js/index.js"></script>
 </body>
 
 </html>
