+++ package.json
@@ -7,11 +7,9 @@
     "example": "examples"
   },
   "scripts": {
-    "build": "npm run build:css && npm run build:js",
-    "build:js": "browserify site/js/index.js -o site/assets/bundle.js -t [ babelify --presets [@babel/preset-env] --plugins [babel-plugin-transform-class-properties @babel/plugin-proposal-object-rest-spread] ]",
-    "build:css": "postcss site/css/styles.css -o site/assets/styles.css",
-    "test": "eslint site/js/*.js",
-    "start": "node server.js --watch"
+    "build": "npx snowpack@2.14.0",
+    "test": "stylelint site/css/*.css && eslint site/js/*.js",
+    "start": "browser-sync --no-notify"
   },
   "repository": {
     "type": "git",
@@ -25,51 +23,31 @@
   "bugs": {
     "url": "https://github.com/sparkbox/bouncy-ball/issues"
   },
-  "homepage": "https://github.com/sparkbox/bouncy-ball#readme",
-  "dependencies": {
-    "platform": "^1.3.5",
-    "prismjs": "^1.17.1",
-    "remarkable": "^2.0.0"
-  },
+  "homepage": "https://github.com/sparkbox/bouncy-ball",
   "devDependencies": {
-    "@babel/cli": "^7.6.4",
-    "@babel/core": "^7.6.4",
-    "@babel/plugin-proposal-object-rest-spread": "^7.6.2",
-    "@babel/preset-env": "^7.6.3",
-    "babel-eslint": "^10.0.3",
-    "babel-plugin-transform-class-properties": "^6.24.1",
-    "babelify": "^10.0.0",
-    "browser-sync": "^2.26.7",
-    "browserify": "^16.5.0",
-    "command-line-args": "^5.1.1",
-    "eslint": "6.6.0",
-    "eslint-config-airbnb": "18.0.1",
-    "eslint-config-sparkbox": "^0.5.6",
-    "eslint-plugin-flowtype": "^4.3.0",
-    "eslint-plugin-import": "^2.18.2",
-    "eslint-plugin-jsx-a11y": "^6.2.3",
-    "eslint-plugin-react": "^7.16.0",
-    "eslint-plugin-react-hooks": "^1.7.0",
-    "http-rewrite-middleware": "^0.1.6",
-    "postcss": "^7.0.21",
-    "postcss-cli": "^6.1.3",
-    "postcss-custom-media": "^7.0.8",
-    "postcss-import": "^12.0.1",
-    "postcss-load-config": "^2.1.0",
-    "postcss-preset-env": "^6.7.0",
-    "postcss-reporter": "^6.0.1",
-    "react": "^16.13.1",
-    "shelljs": "^0.8.3"
+    "@sparkbox/eslint-config-sparkbox": "1.1.5",
+    "@sparkbox/stylelint-config-sparkbox": "^2.0.0",
+    "babel-eslint": "^10.1.0",
+    "browser-sync": "^2.26.13",
+    "eslint": "^7.11.0",
+    "eslint-config-airbnb": "^18.2.0",
+    "eslint-plugin-flowtype": "^5.2.0",
+    "eslint-plugin-import": "^2.22.1",
+    "eslint-plugin-jsx-a11y": "^6.3.1",
+    "eslint-plugin-react": "^7.21.4",
+    "stylelint": "^13.7.2"
   },
-  "postcss": {
-    "plugins": {
-      "postcss-import": {
-        "path": "node_modules/prismjs/themes"
-      },
-      "postcss-preset-env": {
-        "stage": 1
-      },
-      "postcss-reporter": {}
+  "dependencies": {
+    "platform": "^1.3.6",
+    "remarkable": "^2.0.1"
+  },
+  "snowpack": {
+    "install": [
+      "platform",
+      "remarkable"
+    ],
+    "installOptions": {
+      "dest": "site/web_modules"
     }
   }
 }
