+++ site/css/example-colors.css
@@ -1,3 +1,4 @@
+/* stylelint-disable */
 /*
   colors for navigation buttons and bouncy balls
   ! make sure all colors meet WCAG AA standards for normal text
