+++ site/web_modules/remarkable.js
@@ -1,1234 +1,8 @@
-(function(){function r(e,n,t){function o(i,f){if(!n[i]){if(!e[i]){var c="function"==typeof require&&require;if(!f&&c)return c(i,!0);if(u)return u(i,!0);var a=new Error("Cannot find module '"+i+"'");throw a.code="MODULE_NOT_FOUND",a}var p=n[i]={exports:{}};e[i][0].call(p.exports,function(r){var n=e[i][1][r];return o(n||r)},p,p.exports,r,e,n,t)}return n[i].exports}for(var u="function"==typeof require&&require,i=0;i<t.length;i++)o(t[i]);return o}return r})()({1:[function(require,module,exports){
-(function (global){
-/*!
- * Platform.js <https://mths.be/platform>
- * Copyright 2014-2018 Benjamin Tan <https://bnjmnt4n.now.sh/>
- * Copyright 2011-2013 John-David Dalton <http://allyoucanleet.com/>
- * Available under MIT license <https://mths.be/mit>
- */
-;(function() {
-  'use strict';
-
-  /** Used to determine if values are of the language type `Object`. */
-  var objectTypes = {
-    'function': true,
-    'object': true
-  };
-
-  /** Used as a reference to the global object. */
-  var root = (objectTypes[typeof window] && window) || this;
-
-  /** Backup possible global object. */
-  var oldRoot = root;
-
-  /** Detect free variable `exports`. */
-  var freeExports = objectTypes[typeof exports] && exports;
-
-  /** Detect free variable `module`. */
-  var freeModule = objectTypes[typeof module] && module && !module.nodeType && module;
-
-  /** Detect free variable `global` from Node.js or Browserified code and use it as `root`. */
-  var freeGlobal = freeExports && freeModule && typeof global == 'object' && global;
-  if (freeGlobal && (freeGlobal.global === freeGlobal || freeGlobal.window === freeGlobal || freeGlobal.self === freeGlobal)) {
-    root = freeGlobal;
-  }
-
-  /**
-   * Used as the maximum length of an array-like object.
-   * See the [ES6 spec](http://people.mozilla.org/~jorendorff/es6-draft.html#sec-tolength)
-   * for more details.
-   */
-  var maxSafeInteger = Math.pow(2, 53) - 1;
-
-  /** Regular expression to detect Opera. */
-  var reOpera = /\bOpera/;
-
-  /** Possible global object. */
-  var thisBinding = this;
-
-  /** Used for native method references. */
-  var objectProto = Object.prototype;
-
-  /** Used to check for own properties of an object. */
-  var hasOwnProperty = objectProto.hasOwnProperty;
-
-  /** Used to resolve the internal `[[Class]]` of values. */
-  var toString = objectProto.toString;
-
-  /*--------------------------------------------------------------------------*/
-
-  /**
-   * Capitalizes a string value.
-   *
-   * @private
-   * @param {string} string The string to capitalize.
-   * @returns {string} The capitalized string.
-   */
-  function capitalize(string) {
-    string = String(string);
-    return string.charAt(0).toUpperCase() + string.slice(1);
-  }
-
-  /**
-   * A utility function to clean up the OS name.
-   *
-   * @private
-   * @param {string} os The OS name to clean up.
-   * @param {string} [pattern] A `RegExp` pattern matching the OS name.
-   * @param {string} [label] A label for the OS.
-   */
-  function cleanupOS(os, pattern, label) {
-    // Platform tokens are defined at:
-    // http://msdn.microsoft.com/en-us/library/ms537503(VS.85).aspx
-    // http://web.archive.org/web/20081122053950/http://msdn.microsoft.com/en-us/library/ms537503(VS.85).aspx
-    var data = {
-      '10.0': '10',
-      '6.4':  '10 Technical Preview',
-      '6.3':  '8.1',
-      '6.2':  '8',
-      '6.1':  'Server 2008 R2 / 7',
-      '6.0':  'Server 2008 / Vista',
-      '5.2':  'Server 2003 / XP 64-bit',
-      '5.1':  'XP',
-      '5.01': '2000 SP1',
-      '5.0':  '2000',
-      '4.0':  'NT',
-      '4.90': 'ME'
-    };
-    // Detect Windows version from platform tokens.
-    if (pattern && label && /^Win/i.test(os) && !/^Windows Phone /i.test(os) &&
-        (data = data[/[\d.]+$/.exec(os)])) {
-      os = 'Windows ' + data;
-    }
-    // Correct character case and cleanup string.
-    os = String(os);
-
-    if (pattern && label) {
-      os = os.replace(RegExp(pattern, 'i'), label);
-    }
-
-    os = format(
-      os.replace(/ ce$/i, ' CE')
-        .replace(/\bhpw/i, 'web')
-        .replace(/\bMacintosh\b/, 'Mac OS')
-        .replace(/_PowerPC\b/i, ' OS')
-        .replace(/\b(OS X) [^ \d]+/i, '$1')
-        .replace(/\bMac (OS X)\b/, '$1')
-        .replace(/\/(\d)/, ' $1')
-        .replace(/_/g, '.')
-        .replace(/(?: BePC|[ .]*fc[ \d.]+)$/i, '')
-        .replace(/\bx86\.64\b/gi, 'x86_64')
-        .replace(/\b(Windows Phone) OS\b/, '$1')
-        .replace(/\b(Chrome OS \w+) [\d.]+\b/, '$1')
-        .split(' on ')[0]
-    );
-
-    return os;
-  }
-
-  /**
-   * An iteration utility for arrays and objects.
-   *
-   * @private
-   * @param {Array|Object} object The object to iterate over.
-   * @param {Function} callback The function called per iteration.
-   */
-  function each(object, callback) {
-    var index = -1,
-        length = object ? object.length : 0;
-
-    if (typeof length == 'number' && length > -1 && length <= maxSafeInteger) {
-      while (++index < length) {
-        callback(object[index], index, object);
-      }
-    } else {
-      forOwn(object, callback);
-    }
-  }
-
-  /**
-   * Trim and conditionally capitalize string values.
-   *
-   * @private
-   * @param {string} string The string to format.
-   * @returns {string} The formatted string.
-   */
-  function format(string) {
-    string = trim(string);
-    return /^(?:webOS|i(?:OS|P))/.test(string)
-      ? string
-      : capitalize(string);
-  }
-
-  /**
-   * Iterates over an object's own properties, executing the `callback` for each.
-   *
-   * @private
-   * @param {Object} object The object to iterate over.
-   * @param {Function} callback The function executed per own property.
-   */
-  function forOwn(object, callback) {
-    for (var key in object) {
-      if (hasOwnProperty.call(object, key)) {
-        callback(object[key], key, object);
-      }
-    }
-  }
-
-  /**
-   * Gets the internal `[[Class]]` of a value.
-   *
-   * @private
-   * @param {*} value The value.
-   * @returns {string} The `[[Class]]`.
-   */
-  function getClassOf(value) {
-    return value == null
-      ? capitalize(value)
-      : toString.call(value).slice(8, -1);
-  }
-
-  /**
-   * Host objects can return type values that are different from their actual
-   * data type. The objects we are concerned with usually return non-primitive
-   * types of "object", "function", or "unknown".
-   *
-   * @private
-   * @param {*} object The owner of the property.
-   * @param {string} property The property to check.
-   * @returns {boolean} Returns `true` if the property value is a non-primitive, else `false`.
-   */
-  function isHostType(object, property) {
-    var type = object != null ? typeof object[property] : 'number';
-    return !/^(?:boolean|number|string|undefined)$/.test(type) &&
-      (type == 'object' ? !!object[property] : true);
-  }
-
-  /**
-   * Prepares a string for use in a `RegExp` by making hyphens and spaces optional.
-   *
-   * @private
-   * @param {string} string The string to qualify.
-   * @returns {string} The qualified string.
-   */
-  function qualify(string) {
-    return String(string).replace(/([ -])(?!$)/g, '$1?');
-  }
-
-  /**
-   * A bare-bones `Array#reduce` like utility function.
-   *
-   * @private
-   * @param {Array} array The array to iterate over.
-   * @param {Function} callback The function called per iteration.
-   * @returns {*} The accumulated result.
-   */
-  function reduce(array, callback) {
-    var accumulator = null;
-    each(array, function(value, index) {
-      accumulator = callback(accumulator, value, index, array);
-    });
-    return accumulator;
-  }
-
-  /**
-   * Removes leading and trailing whitespace from a string.
-   *
-   * @private
-   * @param {string} string The string to trim.
-   * @returns {string} The trimmed string.
-   */
-  function trim(string) {
-    return String(string).replace(/^ +| +$/g, '');
-  }
-
-  /*--------------------------------------------------------------------------*/
-
-  /**
-   * Creates a new platform object.
-   *
-   * @memberOf platform
-   * @param {Object|string} [ua=navigator.userAgent] The user agent string or
-   *  context object.
-   * @returns {Object} A platform object.
-   */
-  function parse(ua) {
-
-    /** The environment context object. */
-    var context = root;
-
-    /** Used to flag when a custom context is provided. */
-    var isCustomContext = ua && typeof ua == 'object' && getClassOf(ua) != 'String';
-
-    // Juggle arguments.
-    if (isCustomContext) {
-      context = ua;
-      ua = null;
-    }
-
-    /** Browser navigator object. */
-    var nav = context.navigator || {};
-
-    /** Browser user agent string. */
-    var userAgent = nav.userAgent || '';
-
-    ua || (ua = userAgent);
-
-    /** Used to flag when `thisBinding` is the [ModuleScope]. */
-    var isModuleScope = isCustomContext || thisBinding == oldRoot;
-
-    /** Used to detect if browser is like Chrome. */
-    var likeChrome = isCustomContext
-      ? !!nav.likeChrome
-      : /\bChrome\b/.test(ua) && !/internal|\n/i.test(toString.toString());
-
-    /** Internal `[[Class]]` value shortcuts. */
-    var objectClass = 'Object',
-        airRuntimeClass = isCustomContext ? objectClass : 'ScriptBridgingProxyObject',
-        enviroClass = isCustomContext ? objectClass : 'Environment',
-        javaClass = (isCustomContext && context.java) ? 'JavaPackage' : getClassOf(context.java),
-        phantomClass = isCustomContext ? objectClass : 'RuntimeObject';
-
-    /** Detect Java environments. */
-    var java = /\bJava/.test(javaClass) && context.java;
-
-    /** Detect Rhino. */
-    var rhino = java && getClassOf(context.environment) == enviroClass;
-
-    /** A character to represent alpha. */
-    var alpha = java ? 'a' : '\u03b1';
-
-    /** A character to represent beta. */
-    var beta = java ? 'b' : '\u03b2';
-
-    /** Browser document object. */
-    var doc = context.document || {};
-
-    /**
-     * Detect Opera browser (Presto-based).
-     * http://www.howtocreate.co.uk/operaStuff/operaObject.html
-     * http://dev.opera.com/articles/view/opera-mini-web-content-authoring-guidelines/#operamini
-     */
-    var opera = context.operamini || context.opera;
-
-    /** Opera `[[Class]]`. */
-    var operaClass = reOpera.test(operaClass = (isCustomContext && opera) ? opera['[[Class]]'] : getClassOf(opera))
-      ? operaClass
-      : (opera = null);
-
-    /*------------------------------------------------------------------------*/
-
-    /** Temporary variable used over the script's lifetime. */
-    var data;
-
-    /** The CPU architecture. */
-    var arch = ua;
-
-    /** Platform description array. */
-    var description = [];
-
-    /** Platform alpha/beta indicator. */
-    var prerelease = null;
-
-    /** A flag to indicate that environment features should be used to resolve the platform. */
-    var useFeatures = ua == userAgent;
-
-    /** The browser/environment version. */
-    var version = useFeatures && opera && typeof opera.version == 'function' && opera.version();
-
-    /** A flag to indicate if the OS ends with "/ Version" */
-    var isSpecialCasedOS;
-
-    /* Detectable layout engines (order is important). */
-    var layout = getLayout([
-      { 'label': 'EdgeHTML', 'pattern': 'Edge' },
-      'Trident',
-      { 'label': 'WebKit', 'pattern': 'AppleWebKit' },
-      'iCab',
-      'Presto',
-      'NetFront',
-      'Tasman',
-      'KHTML',
-      'Gecko'
-    ]);
-
-    /* Detectable browser names (order is important). */
-    var name = getName([
-      'Adobe AIR',
-      'Arora',
-      'Avant Browser',
-      'Breach',
-      'Camino',
-      'Electron',
-      'Epiphany',
-      'Fennec',
-      'Flock',
-      'Galeon',
-      'GreenBrowser',
-      'iCab',
-      'Iceweasel',
-      'K-Meleon',
-      'Konqueror',
-      'Lunascape',
-      'Maxthon',
-      { 'label': 'Microsoft Edge', 'pattern': 'Edge' },
-      'Midori',
-      'Nook Browser',
-      'PaleMoon',
-      'PhantomJS',
-      'Raven',
-      'Rekonq',
-      'RockMelt',
-      { 'label': 'Samsung Internet', 'pattern': 'SamsungBrowser' },
-      'SeaMonkey',
-      { 'label': 'Silk', 'pattern': '(?:Cloud9|Silk-Accelerated)' },
-      'Sleipnir',
-      'SlimBrowser',
-      { 'label': 'SRWare Iron', 'pattern': 'Iron' },
-      'Sunrise',
-      'Swiftfox',
-      'Waterfox',
-      'WebPositive',
-      'Opera Mini',
-      { 'label': 'Opera Mini', 'pattern': 'OPiOS' },
-      'Opera',
-      { 'label': 'Opera', 'pattern': 'OPR' },
-      'Chrome',
-      { 'label': 'Chrome Mobile', 'pattern': '(?:CriOS|CrMo)' },
-      { 'label': 'Firefox', 'pattern': '(?:Firefox|Minefield)' },
-      { 'label': 'Firefox for iOS', 'pattern': 'FxiOS' },
-      { 'label': 'IE', 'pattern': 'IEMobile' },
-      { 'label': 'IE', 'pattern': 'MSIE' },
-      'Safari'
-    ]);
-
-    /* Detectable products (order is important). */
-    var product = getProduct([
-      { 'label': 'BlackBerry', 'pattern': 'BB10' },
-      'BlackBerry',
-      { 'label': 'Galaxy S', 'pattern': 'GT-I9000' },
-      { 'label': 'Galaxy S2', 'pattern': 'GT-I9100' },
-      { 'label': 'Galaxy S3', 'pattern': 'GT-I9300' },
-      { 'label': 'Galaxy S4', 'pattern': 'GT-I9500' },
-      { 'label': 'Galaxy S5', 'pattern': 'SM-G900' },
-      { 'label': 'Galaxy S6', 'pattern': 'SM-G920' },
-      { 'label': 'Galaxy S6 Edge', 'pattern': 'SM-G925' },
-      { 'label': 'Galaxy S7', 'pattern': 'SM-G930' },
-      { 'label': 'Galaxy S7 Edge', 'pattern': 'SM-G935' },
-      'Google TV',
-      'Lumia',
-      'iPad',
-      'iPod',
-      'iPhone',
-      'Kindle',
-      { 'label': 'Kindle Fire', 'pattern': '(?:Cloud9|Silk-Accelerated)' },
-      'Nexus',
-      'Nook',
-      'PlayBook',
-      'PlayStation Vita',
-      'PlayStation',
-      'TouchPad',
-      'Transformer',
-      { 'label': 'Wii U', 'pattern': 'WiiU' },
-      'Wii',
-      'Xbox One',
-      { 'label': 'Xbox 360', 'pattern': 'Xbox' },
-      'Xoom'
-    ]);
-
-    /* Detectable manufacturers. */
-    var manufacturer = getManufacturer({
-      'Apple': { 'iPad': 1, 'iPhone': 1, 'iPod': 1 },
-      'Archos': {},
-      'Amazon': { 'Kindle': 1, 'Kindle Fire': 1 },
-      'Asus': { 'Transformer': 1 },
-      'Barnes & Noble': { 'Nook': 1 },
-      'BlackBerry': { 'PlayBook': 1 },
-      'Google': { 'Google TV': 1, 'Nexus': 1 },
-      'HP': { 'TouchPad': 1 },
-      'HTC': {},
-      'LG': {},
-      'Microsoft': { 'Xbox': 1, 'Xbox One': 1 },
-      'Motorola': { 'Xoom': 1 },
-      'Nintendo': { 'Wii U': 1,  'Wii': 1 },
-      'Nokia': { 'Lumia': 1 },
-      'Samsung': { 'Galaxy S': 1, 'Galaxy S2': 1, 'Galaxy S3': 1, 'Galaxy S4': 1 },
-      'Sony': { 'PlayStation': 1, 'PlayStation Vita': 1 }
-    });
-
-    /* Detectable operating systems (order is important). */
-    var os = getOS([
-      'Windows Phone',
-      'Android',
-      'CentOS',
-      { 'label': 'Chrome OS', 'pattern': 'CrOS' },
-      'Debian',
-      'Fedora',
-      'FreeBSD',
-      'Gentoo',
-      'Haiku',
-      'Kubuntu',
-      'Linux Mint',
-      'OpenBSD',
-      'Red Hat',
-      'SuSE',
-      'Ubuntu',
-      'Xubuntu',
-      'Cygwin',
-      'Symbian OS',
-      'hpwOS',
-      'webOS ',
-      'webOS',
-      'Tablet OS',
-      'Tizen',
-      'Linux',
-      'Mac OS X',
-      'Macintosh',
-      'Mac',
-      'Windows 98;',
-      'Windows '
-    ]);
-
-    /*------------------------------------------------------------------------*/
-
-    /**
-     * Picks the layout engine from an array of guesses.
-     *
-     * @private
-     * @param {Array} guesses An array of guesses.
-     * @returns {null|string} The detected layout engine.
-     */
-    function getLayout(guesses) {
-      return reduce(guesses, function(result, guess) {
-        return result || RegExp('\\b' + (
-          guess.pattern || qualify(guess)
-        ) + '\\b', 'i').exec(ua) && (guess.label || guess);
-      });
-    }
-
-    /**
-     * Picks the manufacturer from an array of guesses.
-     *
-     * @private
-     * @param {Array} guesses An object of guesses.
-     * @returns {null|string} The detected manufacturer.
-     */
-    function getManufacturer(guesses) {
-      return reduce(guesses, function(result, value, key) {
-        // Lookup the manufacturer by product or scan the UA for the manufacturer.
-        return result || (
-          value[product] ||
-          value[/^[a-z]+(?: +[a-z]+\b)*/i.exec(product)] ||
-          RegExp('\\b' + qualify(key) + '(?:\\b|\\w*\\d)', 'i').exec(ua)
-        ) && key;
-      });
-    }
-
-    /**
-     * Picks the browser name from an array of guesses.
-     *
-     * @private
-     * @param {Array} guesses An array of guesses.
-     * @returns {null|string} The detected browser name.
-     */
-    function getName(guesses) {
-      return reduce(guesses, function(result, guess) {
-        return result || RegExp('\\b' + (
-          guess.pattern || qualify(guess)
-        ) + '\\b', 'i').exec(ua) && (guess.label || guess);
-      });
-    }
-
-    /**
-     * Picks the OS name from an array of guesses.
-     *
-     * @private
-     * @param {Array} guesses An array of guesses.
-     * @returns {null|string} The detected OS name.
-     */
-    function getOS(guesses) {
-      return reduce(guesses, function(result, guess) {
-        var pattern = guess.pattern || qualify(guess);
-        if (!result && (result =
-              RegExp('\\b' + pattern + '(?:/[\\d.]+|[ \\w.]*)', 'i').exec(ua)
-            )) {
-          result = cleanupOS(result, pattern, guess.label || guess);
-        }
-        return result;
-      });
-    }
-
-    /**
-     * Picks the product name from an array of guesses.
-     *
-     * @private
-     * @param {Array} guesses An array of guesses.
-     * @returns {null|string} The detected product name.
-     */
-    function getProduct(guesses) {
-      return reduce(guesses, function(result, guess) {
-        var pattern = guess.pattern || qualify(guess);
-        if (!result && (result =
-              RegExp('\\b' + pattern + ' *\\d+[.\\w_]*', 'i').exec(ua) ||
-              RegExp('\\b' + pattern + ' *\\w+-[\\w]*', 'i').exec(ua) ||
-              RegExp('\\b' + pattern + '(?:; *(?:[a-z]+[_-])?[a-z]+\\d+|[^ ();-]*)', 'i').exec(ua)
-            )) {
-          // Split by forward slash and append product version if needed.
-          if ((result = String((guess.label && !RegExp(pattern, 'i').test(guess.label)) ? guess.label : result).split('/'))[1] && !/[\d.]+/.test(result[0])) {
-            result[0] += ' ' + result[1];
-          }
-          // Correct character case and cleanup string.
-          guess = guess.label || guess;
-          result = format(result[0]
-            .replace(RegExp(pattern, 'i'), guess)
-            .replace(RegExp('; *(?:' + guess + '[_-])?', 'i'), ' ')
-            .replace(RegExp('(' + guess + ')[-_.]?(\\w)', 'i'), '$1 $2'));
-        }
-        return result;
-      });
-    }
-
-    /**
-     * Resolves the version using an array of UA patterns.
-     *
-     * @private
-     * @param {Array} patterns An array of UA patterns.
-     * @returns {null|string} The detected version.
-     */
-    function getVersion(patterns) {
-      return reduce(patterns, function(result, pattern) {
-        return result || (RegExp(pattern +
-          '(?:-[\\d.]+/|(?: for [\\w-]+)?[ /-])([\\d.]+[^ ();/_-]*)', 'i').exec(ua) || 0)[1] || null;
-      });
-    }
-
-    /**
-     * Returns `platform.description` when the platform object is coerced to a string.
-     *
-     * @name toString
-     * @memberOf platform
-     * @returns {string} Returns `platform.description` if available, else an empty string.
-     */
-    function toStringPlatform() {
-      return this.description || '';
-    }
-
-    /*------------------------------------------------------------------------*/
-
-    // Convert layout to an array so we can add extra details.
-    layout && (layout = [layout]);
-
-    // Detect product names that contain their manufacturer's name.
-    if (manufacturer && !product) {
-      product = getProduct([manufacturer]);
-    }
-    // Clean up Google TV.
-    if ((data = /\bGoogle TV\b/.exec(product))) {
-      product = data[0];
-    }
-    // Detect simulators.
-    if (/\bSimulator\b/i.test(ua)) {
-      product = (product ? product + ' ' : '') + 'Simulator';
-    }
-    // Detect Opera Mini 8+ running in Turbo/Uncompressed mode on iOS.
-    if (name == 'Opera Mini' && /\bOPiOS\b/.test(ua)) {
-      description.push('running in Turbo/Uncompressed mode');
-    }
-    // Detect IE Mobile 11.
-    if (name == 'IE' && /\blike iPhone OS\b/.test(ua)) {
-      data = parse(ua.replace(/like iPhone OS/, ''));
-      manufacturer = data.manufacturer;
-      product = data.product;
-    }
-    // Detect iOS.
-    else if (/^iP/.test(product)) {
-      name || (name = 'Safari');
-      os = 'iOS' + ((data = / OS ([\d_]+)/i.exec(ua))
-        ? ' ' + data[1].replace(/_/g, '.')
-        : '');
-    }
-    // Detect Kubuntu.
-    else if (name == 'Konqueror' && !/buntu/i.test(os)) {
-      os = 'Kubuntu';
-    }
-    // Detect Android browsers.
-    else if ((manufacturer && manufacturer != 'Google' &&
-        ((/Chrome/.test(name) && !/\bMobile Safari\b/i.test(ua)) || /\bVita\b/.test(product))) ||
-        (/\bAndroid\b/.test(os) && /^Chrome/.test(name) && /\bVersion\//i.test(ua))) {
-      name = 'Android Browser';
-      os = /\bAndroid\b/.test(os) ? os : 'Android';
-    }
-    // Detect Silk desktop/accelerated modes.
-    else if (name == 'Silk') {
-      if (!/\bMobi/i.test(ua)) {
-        os = 'Android';
-        description.unshift('desktop mode');
-      }
-      if (/Accelerated *= *true/i.test(ua)) {
-        description.unshift('accelerated');
-      }
-    }
-    // Detect PaleMoon identifying as Firefox.
-    else if (name == 'PaleMoon' && (data = /\bFirefox\/([\d.]+)\b/.exec(ua))) {
-      description.push('identifying as Firefox ' + data[1]);
-    }
-    // Detect Firefox OS and products running Firefox.
-    else if (name == 'Firefox' && (data = /\b(Mobile|Tablet|TV)\b/i.exec(ua))) {
-      os || (os = 'Firefox OS');
-      product || (product = data[1]);
-    }
-    // Detect false positives for Firefox/Safari.
-    else if (!name || (data = !/\bMinefield\b/i.test(ua) && /\b(?:Firefox|Safari)\b/.exec(name))) {
-      // Escape the `/` for Firefox 1.
-      if (name && !product && /[\/,]|^[^(]+?\)/.test(ua.slice(ua.indexOf(data + '/') + 8))) {
-        // Clear name of false positives.
-        name = null;
-      }
-      // Reassign a generic name.
-      if ((data = product || manufacturer || os) &&
-          (product || manufacturer || /\b(?:Android|Symbian OS|Tablet OS|webOS)\b/.test(os))) {
-        name = /[a-z]+(?: Hat)?/i.exec(/\bAndroid\b/.test(os) ? os : data) + ' Browser';
-      }
-    }
-    // Add Chrome version to description for Electron.
-    else if (name == 'Electron' && (data = (/\bChrome\/([\d.]+)\b/.exec(ua) || 0)[1])) {
-      description.push('Chromium ' + data);
-    }
-    // Detect non-Opera (Presto-based) versions (order is important).
-    if (!version) {
-      version = getVersion([
-        '(?:Cloud9|CriOS|CrMo|Edge|FxiOS|IEMobile|Iron|Opera ?Mini|OPiOS|OPR|Raven|SamsungBrowser|Silk(?!/[\\d.]+$))',
-        'Version',
-        qualify(name),
-        '(?:Firefox|Minefield|NetFront)'
-      ]);
-    }
-    // Detect stubborn layout engines.
-    if ((data =
-          layout == 'iCab' && parseFloat(version) > 3 && 'WebKit' ||
-          /\bOpera\b/.test(name) && (/\bOPR\b/.test(ua) ? 'Blink' : 'Presto') ||
-          /\b(?:Midori|Nook|Safari)\b/i.test(ua) && !/^(?:Trident|EdgeHTML)$/.test(layout) && 'WebKit' ||
-          !layout && /\bMSIE\b/i.test(ua) && (os == 'Mac OS' ? 'Tasman' : 'Trident') ||
-          layout == 'WebKit' && /\bPlayStation\b(?! Vita\b)/i.test(name) && 'NetFront'
-        )) {
-      layout = [data];
-    }
-    // Detect Windows Phone 7 desktop mode.
-    if (name == 'IE' && (data = (/; *(?:XBLWP|ZuneWP)(\d+)/i.exec(ua) || 0)[1])) {
-      name += ' Mobile';
-      os = 'Windows Phone ' + (/\+$/.test(data) ? data : data + '.x');
-      description.unshift('desktop mode');
-    }
-    // Detect Windows Phone 8.x desktop mode.
-    else if (/\bWPDesktop\b/i.test(ua)) {
-      name = 'IE Mobile';
-      os = 'Windows Phone 8.x';
-      description.unshift('desktop mode');
-      version || (version = (/\brv:([\d.]+)/.exec(ua) || 0)[1]);
-    }
-    // Detect IE 11 identifying as other browsers.
-    else if (name != 'IE' && layout == 'Trident' && (data = /\brv:([\d.]+)/.exec(ua))) {
-      if (name) {
-        description.push('identifying as ' + name + (version ? ' ' + version : ''));
-      }
-      name = 'IE';
-      version = data[1];
-    }
-    // Leverage environment features.
-    if (useFeatures) {
-      // Detect server-side environments.
-      // Rhino has a global function while others have a global object.
-      if (isHostType(context, 'global')) {
-        if (java) {
-          data = java.lang.System;
-          arch = data.getProperty('os.arch');
-          os = os || data.getProperty('os.name') + ' ' + data.getProperty('os.version');
-        }
-        if (rhino) {
-          try {
-            version = context.require('ringo/engine').version.join('.');
-            name = 'RingoJS';
-          } catch(e) {
-            if ((data = context.system) && data.global.system == context.system) {
-              name = 'Narwhal';
-              os || (os = data[0].os || null);
-            }
-          }
-          if (!name) {
-            name = 'Rhino';
-          }
-        }
-        else if (
-          typeof context.process == 'object' && !context.process.browser &&
-          (data = context.process)
-        ) {
-          if (typeof data.versions == 'object') {
-            if (typeof data.versions.electron == 'string') {
-              description.push('Node ' + data.versions.node);
-              name = 'Electron';
-              version = data.versions.electron;
-            } else if (typeof data.versions.nw == 'string') {
-              description.push('Chromium ' + version, 'Node ' + data.versions.node);
-              name = 'NW.js';
-              version = data.versions.nw;
-            }
-          }
-          if (!name) {
-            name = 'Node.js';
-            arch = data.arch;
-            os = data.platform;
-            version = /[\d.]+/.exec(data.version);
-            version = version ? version[0] : null;
-          }
-        }
-      }
-      // Detect Adobe AIR.
-      else if (getClassOf((data = context.runtime)) == airRuntimeClass) {
-        name = 'Adobe AIR';
-        os = data.flash.system.Capabilities.os;
-      }
-      // Detect PhantomJS.
-      else if (getClassOf((data = context.phantom)) == phantomClass) {
-        name = 'PhantomJS';
-        version = (data = data.version || null) && (data.major + '.' + data.minor + '.' + data.patch);
-      }
-      // Detect IE compatibility modes.
-      else if (typeof doc.documentMode == 'number' && (data = /\bTrident\/(\d+)/i.exec(ua))) {
-        // We're in compatibility mode when the Trident version + 4 doesn't
-        // equal the document mode.
-        version = [version, doc.documentMode];
-        if ((data = +data[1] + 4) != version[1]) {
-          description.push('IE ' + version[1] + ' mode');
-          layout && (layout[1] = '');
-          version[1] = data;
-        }
-        version = name == 'IE' ? String(version[1].toFixed(1)) : version[0];
-      }
-      // Detect IE 11 masking as other browsers.
-      else if (typeof doc.documentMode == 'number' && /^(?:Chrome|Firefox)\b/.test(name)) {
-        description.push('masking as ' + name + ' ' + version);
-        name = 'IE';
-        version = '11.0';
-        layout = ['Trident'];
-        os = 'Windows';
-      }
-      os = os && format(os);
-    }
-    // Detect prerelease phases.
-    if (version && (data =
-          /(?:[ab]|dp|pre|[ab]\d+pre)(?:\d+\+?)?$/i.exec(version) ||
-          /(?:alpha|beta)(?: ?\d)?/i.exec(ua + ';' + (useFeatures && nav.appMinorVersion)) ||
-          /\bMinefield\b/i.test(ua) && 'a'
-        )) {
-      prerelease = /b/i.test(data) ? 'beta' : 'alpha';
-      version = version.replace(RegExp(data + '\\+?$'), '') +
-        (prerelease == 'beta' ? beta : alpha) + (/\d+\+?/.exec(data) || '');
-    }
-    // Detect Firefox Mobile.
-    if (name == 'Fennec' || name == 'Firefox' && /\b(?:Android|Firefox OS)\b/.test(os)) {
-      name = 'Firefox Mobile';
-    }
-    // Obscure Maxthon's unreliable version.
-    else if (name == 'Maxthon' && version) {
-      version = version.replace(/\.[\d.]+/, '.x');
-    }
-    // Detect Xbox 360 and Xbox One.
-    else if (/\bXbox\b/i.test(product)) {
-      if (product == 'Xbox 360') {
-        os = null;
-      }
-      if (product == 'Xbox 360' && /\bIEMobile\b/.test(ua)) {
-        description.unshift('mobile mode');
-      }
-    }
-    // Add mobile postfix.
-    else if ((/^(?:Chrome|IE|Opera)$/.test(name) || name && !product && !/Browser|Mobi/.test(name)) &&
-        (os == 'Windows CE' || /Mobi/i.test(ua))) {
-      name += ' Mobile';
-    }
-    // Detect IE platform preview.
-    else if (name == 'IE' && useFeatures) {
-      try {
-        if (context.external === null) {
-          description.unshift('platform preview');
-        }
-      } catch(e) {
-        description.unshift('embedded');
-      }
-    }
-    // Detect BlackBerry OS version.
-    // http://docs.blackberry.com/en/developers/deliverables/18169/HTTP_headers_sent_by_BB_Browser_1234911_11.jsp
-    else if ((/\bBlackBerry\b/.test(product) || /\bBB10\b/.test(ua)) && (data =
-          (RegExp(product.replace(/ +/g, ' *') + '/([.\\d]+)', 'i').exec(ua) || 0)[1] ||
-          version
-        )) {
-      data = [data, /BB10/.test(ua)];
-      os = (data[1] ? (product = null, manufacturer = 'BlackBerry') : 'Device Software') + ' ' + data[0];
-      version = null;
-    }
-    // Detect Opera identifying/masking itself as another browser.
-    // http://www.opera.com/support/kb/view/843/
-    else if (this != forOwn && product != 'Wii' && (
-          (useFeatures && opera) ||
-          (/Opera/.test(name) && /\b(?:MSIE|Firefox)\b/i.test(ua)) ||
-          (name == 'Firefox' && /\bOS X (?:\d+\.){2,}/.test(os)) ||
-          (name == 'IE' && (
-            (os && !/^Win/.test(os) && version > 5.5) ||
-            /\bWindows XP\b/.test(os) && version > 8 ||
-            version == 8 && !/\bTrident\b/.test(ua)
-          ))
-        ) && !reOpera.test((data = parse.call(forOwn, ua.replace(reOpera, '') + ';'))) && data.name) {
-      // When "identifying", the UA contains both Opera and the other browser's name.
-      data = 'ing as ' + data.name + ((data = data.version) ? ' ' + data : '');
-      if (reOpera.test(name)) {
-        if (/\bIE\b/.test(data) && os == 'Mac OS') {
-          os = null;
-        }
-        data = 'identify' + data;
-      }
-      // When "masking", the UA contains only the other browser's name.
-      else {
-        data = 'mask' + data;
-        if (operaClass) {
-          name = format(operaClass.replace(/([a-z])([A-Z])/g, '$1 $2'));
-        } else {
-          name = 'Opera';
-        }
-        if (/\bIE\b/.test(data)) {
-          os = null;
-        }
-        if (!useFeatures) {
-          version = null;
-        }
-      }
-      layout = ['Presto'];
-      description.push(data);
-    }
-    // Detect WebKit Nightly and approximate Chrome/Safari versions.
-    if ((data = (/\bAppleWebKit\/([\d.]+\+?)/i.exec(ua) || 0)[1])) {
-      // Correct build number for numeric comparison.
-      // (e.g. "532.5" becomes "532.05")
-      data = [parseFloat(data.replace(/\.(\d)$/, '.0$1')), data];
-      // Nightly builds are postfixed with a "+".
-      if (name == 'Safari' && data[1].slice(-1) == '+') {
-        name = 'WebKit Nightly';
-        prerelease = 'alpha';
-        version = data[1].slice(0, -1);
-      }
-      // Clear incorrect browser versions.
-      else if (version == data[1] ||
-          version == (data[2] = (/\bSafari\/([\d.]+\+?)/i.exec(ua) || 0)[1])) {
-        version = null;
-      }
-      // Use the full Chrome version when available.
-      data[1] = (/\bChrome\/([\d.]+)/i.exec(ua) || 0)[1];
-      // Detect Blink layout engine.
-      if (data[0] == 537.36 && data[2] == 537.36 && parseFloat(data[1]) >= 28 && layout == 'WebKit') {
-        layout = ['Blink'];
-      }
-      // Detect JavaScriptCore.
-      // http://stackoverflow.com/questions/6768474/how-can-i-detect-which-javascript-engine-v8-or-jsc-is-used-at-runtime-in-androi
-      if (!useFeatures || (!likeChrome && !data[1])) {
-        layout && (layout[1] = 'like Safari');
-        data = (data = data[0], data < 400 ? 1 : data < 500 ? 2 : data < 526 ? 3 : data < 533 ? 4 : data < 534 ? '4+' : data < 535 ? 5 : data < 537 ? 6 : data < 538 ? 7 : data < 601 ? 8 : '8');
-      } else {
-        layout && (layout[1] = 'like Chrome');
-        data = data[1] || (data = data[0], data < 530 ? 1 : data < 532 ? 2 : data < 532.05 ? 3 : data < 533 ? 4 : data < 534.03 ? 5 : data < 534.07 ? 6 : data < 534.10 ? 7 : data < 534.13 ? 8 : data < 534.16 ? 9 : data < 534.24 ? 10 : data < 534.30 ? 11 : data < 535.01 ? 12 : data < 535.02 ? '13+' : data < 535.07 ? 15 : data < 535.11 ? 16 : data < 535.19 ? 17 : data < 536.05 ? 18 : data < 536.10 ? 19 : data < 537.01 ? 20 : data < 537.11 ? '21+' : data < 537.13 ? 23 : data < 537.18 ? 24 : data < 537.24 ? 25 : data < 537.36 ? 26 : layout != 'Blink' ? '27' : '28');
-      }
-      // Add the postfix of ".x" or "+" for approximate versions.
-      layout && (layout[1] += ' ' + (data += typeof data == 'number' ? '.x' : /[.+]/.test(data) ? '' : '+'));
-      // Obscure version for some Safari 1-2 releases.
-      if (name == 'Safari' && (!version || parseInt(version) > 45)) {
-        version = data;
-      }
-    }
-    // Detect Opera desktop modes.
-    if (name == 'Opera' &&  (data = /\bzbov|zvav$/.exec(os))) {
-      name += ' ';
-      description.unshift('desktop mode');
-      if (data == 'zvav') {
-        name += 'Mini';
-        version = null;
-      } else {
-        name += 'Mobile';
-      }
-      os = os.replace(RegExp(' *' + data + '$'), '');
-    }
-    // Detect Chrome desktop mode.
-    else if (name == 'Safari' && /\bChrome\b/.exec(layout && layout[1])) {
-      description.unshift('desktop mode');
-      name = 'Chrome Mobile';
-      version = null;
-
-      if (/\bOS X\b/.test(os)) {
-        manufacturer = 'Apple';
-        os = 'iOS 4.3+';
-      } else {
-        os = null;
-      }
-    }
-    // Strip incorrect OS versions.
-    if (version && version.indexOf((data = /[\d.]+$/.exec(os))) == 0 &&
-        ua.indexOf('/' + data + '-') > -1) {
-      os = trim(os.replace(data, ''));
-    }
-    // Add layout engine.
-    if (layout && !/\b(?:Avant|Nook)\b/.test(name) && (
-        /Browser|Lunascape|Maxthon/.test(name) ||
-        name != 'Safari' && /^iOS/.test(os) && /\bSafari\b/.test(layout[1]) ||
-        /^(?:Adobe|Arora|Breach|Midori|Opera|Phantom|Rekonq|Rock|Samsung Internet|Sleipnir|Web)/.test(name) && layout[1])) {
-      // Don't add layout details to description if they are falsey.
-      (data = layout[layout.length - 1]) && description.push(data);
-    }
-    // Combine contextual information.
-    if (description.length) {
-      description = ['(' + description.join('; ') + ')'];
-    }
-    // Append manufacturer to description.
-    if (manufacturer && product && product.indexOf(manufacturer) < 0) {
-      description.push('on ' + manufacturer);
-    }
-    // Append product to description.
-    if (product) {
-      description.push((/^on /.test(description[description.length - 1]) ? '' : 'on ') + product);
-    }
-    // Parse the OS into an object.
-    if (os) {
-      data = / ([\d.+]+)$/.exec(os);
-      isSpecialCasedOS = data && os.charAt(os.length - data[0].length - 1) == '/';
-      os = {
-        'architecture': 32,
-        'family': (data && !isSpecialCasedOS) ? os.replace(data[0], '') : os,
-        'version': data ? data[1] : null,
-        'toString': function() {
-          var version = this.version;
-          return this.family + ((version && !isSpecialCasedOS) ? ' ' + version : '') + (this.architecture == 64 ? ' 64-bit' : '');
-        }
-      };
-    }
-    // Add browser/OS architecture.
-    if ((data = /\b(?:AMD|IA|Win|WOW|x86_|x)64\b/i.exec(arch)) && !/\bi686\b/i.test(arch)) {
-      if (os) {
-        os.architecture = 64;
-        os.family = os.family.replace(RegExp(' *' + data), '');
-      }
-      if (
-          name && (/\bWOW64\b/i.test(ua) ||
-          (useFeatures && /\w(?:86|32)$/.test(nav.cpuClass || nav.platform) && !/\bWin64; x64\b/i.test(ua)))
-      ) {
-        description.unshift('32-bit');
-      }
-    }
-    // Chrome 39 and above on OS X is always 64-bit.
-    else if (
-        os && /^OS X/.test(os.family) &&
-        name == 'Chrome' && parseFloat(version) >= 39
-    ) {
-      os.architecture = 64;
-    }
-
-    ua || (ua = null);
-
-    /*------------------------------------------------------------------------*/
-
-    /**
-     * The platform object.
-     *
-     * @name platform
-     * @type Object
-     */
-    var platform = {};
-
-    /**
-     * The platform description.
-     *
-     * @memberOf platform
-     * @type string|null
-     */
-    platform.description = ua;
-
-    /**
-     * The name of the browser's layout engine.
-     *
-     * The list of common layout engines include:
-     * "Blink", "EdgeHTML", "Gecko", "Trident" and "WebKit"
-     *
-     * @memberOf platform
-     * @type string|null
-     */
-    platform.layout = layout && layout[0];
-
-    /**
-     * The name of the product's manufacturer.
-     *
-     * The list of manufacturers include:
-     * "Apple", "Archos", "Amazon", "Asus", "Barnes & Noble", "BlackBerry",
-     * "Google", "HP", "HTC", "LG", "Microsoft", "Motorola", "Nintendo",
-     * "Nokia", "Samsung" and "Sony"
-     *
-     * @memberOf platform
-     * @type string|null
-     */
-    platform.manufacturer = manufacturer;
-
-    /**
-     * The name of the browser/environment.
-     *
-     * The list of common browser names include:
-     * "Chrome", "Electron", "Firefox", "Firefox for iOS", "IE",
-     * "Microsoft Edge", "PhantomJS", "Safari", "SeaMonkey", "Silk",
-     * "Opera Mini" and "Opera"
-     *
-     * Mobile versions of some browsers have "Mobile" appended to their name:
-     * eg. "Chrome Mobile", "Firefox Mobile", "IE Mobile" and "Opera Mobile"
-     *
-     * @memberOf platform
-     * @type string|null
-     */
-    platform.name = name;
-
-    /**
-     * The alpha/beta release indicator.
-     *
-     * @memberOf platform
-     * @type string|null
-     */
-    platform.prerelease = prerelease;
-
-    /**
-     * The name of the product hosting the browser.
-     *
-     * The list of common products include:
-     *
-     * "BlackBerry", "Galaxy S4", "Lumia", "iPad", "iPod", "iPhone", "Kindle",
-     * "Kindle Fire", "Nexus", "Nook", "PlayBook", "TouchPad" and "Transformer"
-     *
-     * @memberOf platform
-     * @type string|null
-     */
-    platform.product = product;
-
-    /**
-     * The browser's user agent string.
-     *
-     * @memberOf platform
-     * @type string|null
-     */
-    platform.ua = ua;
-
-    /**
-     * The browser/environment version.
-     *
-     * @memberOf platform
-     * @type string|null
-     */
-    platform.version = name && version;
-
-    /**
-     * The name of the operating system.
-     *
-     * @memberOf platform
-     * @type Object
-     */
-    platform.os = os || {
-
-      /**
-       * The CPU architecture the OS is built for.
-       *
-       * @memberOf platform.os
-       * @type number|null
-       */
-      'architecture': null,
-
-      /**
-       * The family of the OS.
-       *
-       * Common values include:
-       * "Windows", "Windows Server 2008 R2 / 7", "Windows Server 2008 / Vista",
-       * "Windows XP", "OS X", "Ubuntu", "Debian", "Fedora", "Red Hat", "SuSE",
-       * "Android", "iOS" and "Windows Phone"
-       *
-       * @memberOf platform.os
-       * @type string|null
-       */
-      'family': null,
-
-      /**
-       * The version of the OS.
-       *
-       * @memberOf platform.os
-       * @type string|null
-       */
-      'version': null,
-
-      /**
-       * Returns the OS string.
-       *
-       * @memberOf platform.os
-       * @returns {string} The OS string.
-       */
-      'toString': function() { return 'null'; }
-    };
-
-    platform.parse = parse;
-    platform.toString = toStringPlatform;
-
-    if (platform.version) {
-      description.unshift(version);
-    }
-    if (platform.name) {
-      description.unshift(name);
-    }
-    if (os && name && !(os == String(os).split(' ')[0] && (os == name.split(' ')[0] || product))) {
-      description.push(product ? '(' + os + ')' : 'on ' + os);
-    }
-    if (description.length) {
-      platform.description = description.join(' ');
-    }
-    return platform;
-  }
-
-  /*--------------------------------------------------------------------------*/
-
-  // Export platform.
-  var platform = parse();
-
-  // Some AMD build optimizers, like r.js, check for condition patterns like the following:
-  if (typeof define == 'function' && typeof define.amd == 'object' && define.amd) {
-    // Expose platform on the global object to prevent errors when platform is
-    // loaded by a script tag in the presence of an AMD loader.
-    // See http://requirejs.org/docs/errors.html#mismatch for more details.
-    root.platform = platform;
-
-    // Define as an anonymous module so platform can be aliased through path mapping.
-    define(function() {
-      return platform;
-    });
-  }
-  // Check for `exports` after `define` in case a build optimizer adds an `exports` object.
-  else if (freeExports && freeModule) {
-    // Export for CommonJS support.
-    forOwn(platform, function(value, key) {
-      freeExports[key] = value;
-    });
-  }
-  else {
-    // Export to the global object.
-    root.platform = platform;
-  }
-}.call(this));
-
-}).call(this,typeof global !== "undefined" ? global : typeof self !== "undefined" ? self : typeof window !== "undefined" ? window : {})
-},{}],2:[function(require,module,exports){
-'use strict';
-
-Object.defineProperty(exports, '__esModule', { value: true });
-
 var textarea;
 
 function decodeEntity(name) {
   textarea = textarea || document.createElement('textarea');
-  textarea.innerHTML = '&' + name;
+  textarea.innerHTML = '&' + name + ';';
   return textarea.value;
 }
 
@@ -6387,856 +5161,4 @@ Remarkable.prototype.renderInline = function (str, env) {
   return this.renderer.render(this.parseInline(str, env), this.options, env);
 };
 
-exports.Remarkable = Remarkable;
-exports.utils = utils;
-
-},{}],3:[function(require,module,exports){
-"use strict";
-
-var updatePanes = require('./updatePanes');
-
-var toggleDocs = require('./toggleDocs');
-
-function init() {
-  // DOM queries and a URL lookup.
-  var options = document.querySelectorAll('.selection-bar .nav-button');
-  var unsupportedLink = document.querySelector('.unsupported-details');
-  var docsToggleLink = document.querySelector('.docs-toggle-link');
-  var hashOption = window.location.hash; // Pre-select an option, if it is found in the URL fragment.
-
-  if (hashOption) {
-    document.querySelector('.is-active').classList.remove('is-active');
-    document.getElementById(hashOption.slice(1)).classList.add('is-active');
-  }
-
-  updatePanes(); // Set listeners for future updates:
-
-  options.forEach(function (option) {
-    option.addEventListener('click', updatePanes);
-  });
-  docsToggleLink.addEventListener('click', toggleDocs);
-  unsupportedLink.addEventListener('click', toggleDocs);
-}
-
-module.exports = {
-  init: init
-};
-
-},{"./toggleDocs":6,"./updatePanes":7}],4:[function(require,module,exports){
-"use strict";
-
-// Keeping Modernizr here, since I'd like it to be globally accessible.
-//   eslint-disable-next-line no-unused-vars
-var Modernizr = require('./vendor/modernizr-custom');
-
-var App = require('./app'); // outline.js replacement that handles outline styling in our stylesheets.
-
-
-document.addEventListener('mousedown', function () {
-  return document.body.classList.add('no-focus');
-});
-document.addEventListener('keydown', function () {
-  return document.body.classList.remove('no-focus');
-});
-App.init();
-
-},{"./app":3,"./vendor/modernizr-custom":8}],5:[function(require,module,exports){
-"use strict";
-
-function sourceDump(url, dumpLocation, options) {
-  var request = new XMLHttpRequest();
-  request.open('GET', url, true);
-
-  request.onload = function () {
-    if (request.status >= 200 && request.status < 400) {
-      var dumpElement = typeof dumpLocation === 'string' || dumpLocation instanceof String ? document.querySelector(dumpLocation) : dumpLocation; // If no dumpLocation value is provided, it won't dump the text,
-      // and will defer to the successCallback to handle the response.
-
-      if (dumpElement) {
-        dumpElement.textContent = request.responseText;
-      }
-
-      if (options.successCallback) {
-        options.successCallback(request.response);
-      }
-    } else {
-      console.error(request.statusText, Error(request.statusText));
-    }
-  };
-
-  request.onerror = function () {
-    console.error('Request Failed :(', Error('Network Error'));
-
-    if (options.failureCallback) {
-      options.failureCallback();
-    }
-  };
-
-  request.send();
-}
-
-module.exports = sourceDump;
-
-},{}],6:[function(require,module,exports){
-"use strict";
-
-var docsPane = document.querySelector('.docs-pane');
-var docsToggle = document.querySelector('.docs-toggle-link');
-/**
- * Function that toggles the docs pane.
- * @param    Object e - The event object
- */
-
-function toggleDocs(e) {
-  docsPane.classList.toggle('docs-pane_is-open');
-  docsToggle.classList.toggle('docs-toggle-link_is-less');
-  e.preventDefault();
-}
-
-module.exports = toggleDocs;
-
-},{}],7:[function(require,module,exports){
-"use strict";
-
-/*  eslint-disable indent */
-//  I'm disabling this rule here because the patterns and
-//  indenting in this case makes it pretty readable.
-var _require = require('remarkable'),
-    Remarkable = _require.Remarkable;
-
-var Platform = require('platform');
-
-var sourceDump = require('./sourceDump');
-
-var Prism = require('./vendor/prismjs-custom'); // DOM queries
-
-
-var srcPreEl = document.querySelector('.source-pane > pre');
-var srcCodeEl = document.querySelector('.source-pane > pre > code');
-var demoEl = document.querySelector('.demo-frame');
-var docsEl = document.querySelector('.docs-pane-content');
-var docsLinkDemoName = document.querySelector('.demo-name');
-var unsupportedEl = document.querySelector('.unsupported'); // We pull this value to the top level, so callbacks can access its latest value.
-
-var selected;
-/**
- * @private
- */
-
-function showIncompatibilityMessage() {
-  // hide iframe
-  demoEl.style.display = 'none'; // show message
-
-  unsupportedEl.style.display = '';
-}
-/**
- * @private
- */
-
-
-function resetIncompatibilityMessage() {
-  // show iframe
-  demoEl.style.display = ''; // hide message
-
-  unsupportedEl.style.display = 'none';
-}
-/**
- * Runs Remarkable on readme text, and drops it into the docs pane.
- * @private
- */
-
-
-function markdownToHtml(response) {
-  var parser = new Remarkable('commonmark');
-  docsEl.innerHTML = parser.render(response);
-}
-/**
- * Checks if the selected demo is compatible with this browser.
- * @private
- */
-
-
-function isCompatible(selectedId) {
-  var browser = Platform.name;
-
-  if (selectedId === 'smil') {
-    // only return true if there's a Modernizr 👍 and the browser isn't Safari.
-    return Modernizr.smil && browser !== 'Safari';
-  }
-
-  if (selectedId === 'flash') {
-    return browser !== 'Safari';
-  }
-
-  if (selectedId === 'p5') {
-    return Modernizr.webgl;
-  }
-
-  return true;
-}
-/**
- * Runs PrismJS on the page. Designed to be called once the new source is on the page.
- * @private
- */
-
-
-function highlightSource() {
-  var srcLanguage = selected.id === 'css' ? 'css' : selected.id === 'css-step' ? 'css' : selected.id === 'smil' ? 'markup' : selected.id === 'video' ? 'markup' : selected.id === 'flash' ? 'markup' : selected.id === 'animated-gif' ? 'markup' : 'javascript';
-  srcCodeEl.className = '';
-  srcCodeEl.classList.add("language-".concat(srcLanguage));
-  Prism.highlightAll();
-}
-/**
- * Updates the preview & source panes based to match the currently selected option.
- */
-
-
-function updatePanes(event) {
-  if (event && event.type === 'click') {
-    // If a click triggered this function, we'll need to change .is-active
-    // and update the url (these don't need to be done when this function
-    // is run on the initial page load).
-    document.querySelector('.is-active').classList.remove('is-active');
-    event.currentTarget.classList.add('is-active');
-    window.location.hash = selected.id;
-  }
-
-  selected = document.querySelector('.is-active');
-  var name = selected.textContent;
-  var srcFileName = selected.id === 'css' ? 'styles.css' : selected.id === 'css-step' ? 'styles.css' : selected.id === 'smil' ? 'image.svg' : selected.id === 'video' ? 'index.html' : selected.id === 'flash' ? 'index.html' : selected.id === 'animated-gif' ? 'index.html' : 'script.js';
-  var demoFileName = selected.id === 'smil' ? 'image.svg' : 'index.html'; // pane content urls
-
-  var srcUrl = "examples/".concat(selected.id, "/").concat(srcFileName);
-  var demoUrl = "examples/".concat(selected.id, "/").concat(demoFileName);
-  var docsUrl = "examples/".concat(selected.id, "/readme.md"); // Update the source pane (scroll it to the top, and get the new source).
-
-  srcPreEl.scrollTop = 0;
-  sourceDump(srcUrl, srcCodeEl, {
-    successCallback: highlightSource
-  }); // Update the demo pane.
-
-  resetIncompatibilityMessage();
-
-  if (!isCompatible(selected.id)) {
-    showIncompatibilityMessage();
-  } else {
-    demoEl.setAttribute('src', demoUrl);
-  } // Update the docs pane.
-
-
-  docsLinkDemoName.textContent = name;
-  sourceDump(docsUrl, undefined, {
-    successCallback: markdownToHtml
-  });
-}
-
-module.exports = updatePanes;
-
-},{"./sourceDump":5,"./vendor/prismjs-custom":9,"platform":1,"remarkable":2}],8:[function(require,module,exports){
-"use strict";
-
-function _typeof(obj) { if (typeof Symbol === "function" && typeof Symbol.iterator === "symbol") { _typeof = function _typeof(obj) { return typeof obj; }; } else { _typeof = function _typeof(obj) { return obj && typeof Symbol === "function" && obj.constructor === Symbol && obj !== Symbol.prototype ? "symbol" : typeof obj; }; } return _typeof(obj); }
-
-/*! modernizr 3.3.1 (Custom Build) | MIT *
- * https://modernizr.com/download/?-smil-webgl-setclasses !*/
-!function (e, n, t) {
-  function a(e, n) {
-    return _typeof(e) === n;
-  }
-
-  function s() {
-    var e, n, t, s, o, i, c;
-
-    for (var f in l) {
-      if (l.hasOwnProperty(f)) {
-        if (e = [], n = l[f], n.name && (e.push(n.name.toLowerCase()), n.options && n.options.aliases && n.options.aliases.length)) for (t = 0; t < n.options.aliases.length; t++) {
-          e.push(n.options.aliases[t].toLowerCase());
-        }
-
-        for (s = a(n.fn, "function") ? n.fn() : n.fn, o = 0; o < e.length; o++) {
-          i = e[o], c = i.split("."), 1 === c.length ? Modernizr[c[0]] = s : (!Modernizr[c[0]] || Modernizr[c[0]] instanceof Boolean || (Modernizr[c[0]] = new Boolean(Modernizr[c[0]])), Modernizr[c[0]][c[1]] = s), r.push((s ? "" : "no-") + c.join("-"));
-        }
-      }
-    }
-  }
-
-  function o(e) {
-    var n = f.className,
-        t = Modernizr._config.classPrefix || "";
-
-    if (u && (n = n.baseVal), Modernizr._config.enableJSClass) {
-      var a = new RegExp("(^|\\s)" + t + "no-js(\\s|$)");
-      n = n.replace(a, "$1" + t + "js$2");
-    }
-
-    Modernizr._config.enableClasses && (n += " " + t + e.join(" " + t), u ? f.className.baseVal = n : f.className = n);
-  }
-
-  function i() {
-    return "function" != typeof n.createElement ? n.createElement(arguments[0]) : u ? n.createElementNS.call(n, "http://www.w3.org/2000/svg", arguments[0]) : n.createElement.apply(n, arguments);
-  }
-
-  var r = [],
-      l = [],
-      c = {
-    _version: "3.3.1",
-    _config: {
-      classPrefix: "",
-      enableClasses: !0,
-      enableJSClass: !0,
-      usePrefixes: !0
-    },
-    _q: [],
-    on: function on(e, n) {
-      var t = this;
-      setTimeout(function () {
-        n(t[e]);
-      }, 0);
-    },
-    addTest: function addTest(e, n, t) {
-      l.push({
-        name: e,
-        fn: n,
-        options: t
-      });
-    },
-    addAsyncTest: function addAsyncTest(e) {
-      l.push({
-        name: null,
-        fn: e
-      });
-    }
-  },
-      Modernizr = function Modernizr() {};
-
-  Modernizr.prototype = c, Modernizr = new Modernizr();
-  var f = n.documentElement,
-      u = "svg" === f.nodeName.toLowerCase();
-  Modernizr.addTest("webgl", function () {
-    var n = i("canvas"),
-        t = "probablySupportsContext" in n ? "probablySupportsContext" : "supportsContext";
-    return t in n ? n[t]("webgl") || n[t]("experimental-webgl") : "WebGLRenderingContext" in e;
-  });
-  var p = {}.toString;
-  Modernizr.addTest("smil", function () {
-    return !!n.createElementNS && /SVGAnimate/.test(p.call(n.createElementNS("http://www.w3.org/2000/svg", "animate")));
-  }), s(), o(r), delete c.addTest, delete c.addAsyncTest;
-
-  for (var m = 0; m < Modernizr._q.length; m++) {
-    Modernizr._q[m]();
-  }
-
-  e.Modernizr = Modernizr;
-}(window, document);
-
-},{}],9:[function(require,module,exports){
-(function (global){
-"use strict";
-
-/* http://prismjs.com/download.html?themes=prism&languages=markup+css+clike+javascript&plugins=autolinker+show-language */
-var _self = "undefined" != typeof window ? window : "undefined" != typeof WorkerGlobalScope && self instanceof WorkerGlobalScope ? self : {},
-    Prism = function () {
-  var e = /\blang(?:uage)?-(\w+)\b/i,
-      t = 0,
-      n = _self.Prism = {
-    util: {
-      encode: function encode(e) {
-        return e instanceof a ? new a(e.type, n.util.encode(e.content), e.alias) : "Array" === n.util.type(e) ? e.map(n.util.encode) : e.replace(/&/g, "&amp;").replace(/</g, "&lt;").replace(/\u00a0/g, " ");
-      },
-      type: function type(e) {
-        return Object.prototype.toString.call(e).match(/\[object (\w+)\]/)[1];
-      },
-      objId: function objId(e) {
-        return e.__id || Object.defineProperty(e, "__id", {
-          value: ++t
-        }), e.__id;
-      },
-      clone: function clone(e) {
-        var t = n.util.type(e);
-
-        switch (t) {
-          case "Object":
-            var a = {};
-
-            for (var r in e) {
-              e.hasOwnProperty(r) && (a[r] = n.util.clone(e[r]));
-            }
-
-            return a;
-
-          case "Array":
-            return e.map && e.map(function (e) {
-              return n.util.clone(e);
-            });
-        }
-
-        return e;
-      }
-    },
-    languages: {
-      extend: function extend(e, t) {
-        var a = n.util.clone(n.languages[e]);
-
-        for (var r in t) {
-          a[r] = t[r];
-        }
-
-        return a;
-      },
-      insertBefore: function insertBefore(e, t, a, r) {
-        r = r || n.languages;
-        var i = r[e];
-
-        if (2 == arguments.length) {
-          a = arguments[1];
-
-          for (var l in a) {
-            a.hasOwnProperty(l) && (i[l] = a[l]);
-          }
-
-          return i;
-        }
-
-        var o = {};
-
-        for (var s in i) {
-          if (i.hasOwnProperty(s)) {
-            if (s == t) for (var l in a) {
-              a.hasOwnProperty(l) && (o[l] = a[l]);
-            }
-            o[s] = i[s];
-          }
-        }
-
-        return n.languages.DFS(n.languages, function (t, n) {
-          n === r[e] && t != e && (this[t] = o);
-        }), r[e] = o;
-      },
-      DFS: function DFS(e, t, a, r) {
-        r = r || {};
-
-        for (var i in e) {
-          e.hasOwnProperty(i) && (t.call(e, i, e[i], a || i), "Object" !== n.util.type(e[i]) || r[n.util.objId(e[i])] ? "Array" !== n.util.type(e[i]) || r[n.util.objId(e[i])] || (r[n.util.objId(e[i])] = !0, n.languages.DFS(e[i], t, i, r)) : (r[n.util.objId(e[i])] = !0, n.languages.DFS(e[i], t, null, r)));
-        }
-      }
-    },
-    plugins: {},
-    highlightAll: function highlightAll(e, t) {
-      var a = {
-        callback: t,
-        selector: 'code[class*="language-"], [class*="language-"] code, code[class*="lang-"], [class*="lang-"] code'
-      };
-      n.hooks.run("before-highlightall", a);
-
-      for (var r, i = a.elements || document.querySelectorAll(a.selector), l = 0; r = i[l++];) {
-        n.highlightElement(r, e === !0, a.callback);
-      }
-    },
-    highlightElement: function highlightElement(t, a, r) {
-      for (var i, l, o = t; o && !e.test(o.className);) {
-        o = o.parentNode;
-      }
-
-      o && (i = (o.className.match(e) || [, ""])[1].toLowerCase(), l = n.languages[i]), t.className = t.className.replace(e, "").replace(/\s+/g, " ") + " language-" + i, o = t.parentNode, /pre/i.test(o.nodeName) && (o.className = o.className.replace(e, "").replace(/\s+/g, " ") + " language-" + i);
-      var s = t.textContent,
-          u = {
-        element: t,
-        language: i,
-        grammar: l,
-        code: s
-      };
-      if (n.hooks.run("before-sanity-check", u), !u.code || !u.grammar) return n.hooks.run("complete", u), void 0;
-
-      if (n.hooks.run("before-highlight", u), a && _self.Worker) {
-        var c = new Worker(n.filename);
-        c.onmessage = function (e) {
-          u.highlightedCode = e.data, n.hooks.run("before-insert", u), u.element.innerHTML = u.highlightedCode, r && r.call(u.element), n.hooks.run("after-highlight", u), n.hooks.run("complete", u);
-        }, c.postMessage(JSON.stringify({
-          language: u.language,
-          code: u.code,
-          immediateClose: !0
-        }));
-      } else u.highlightedCode = n.highlight(u.code, u.grammar, u.language), n.hooks.run("before-insert", u), u.element.innerHTML = u.highlightedCode, r && r.call(t), n.hooks.run("after-highlight", u), n.hooks.run("complete", u);
-    },
-    highlight: function highlight(e, t, r) {
-      var i = n.tokenize(e, t);
-      return a.stringify(n.util.encode(i), r);
-    },
-    tokenize: function tokenize(e, t) {
-      var a = n.Token,
-          r = [e],
-          i = t.rest;
-
-      if (i) {
-        for (var l in i) {
-          t[l] = i[l];
-        }
-
-        delete t.rest;
-      }
-
-      e: for (var l in t) {
-        if (t.hasOwnProperty(l) && t[l]) {
-          var o = t[l];
-          o = "Array" === n.util.type(o) ? o : [o];
-
-          for (var s = 0; s < o.length; ++s) {
-            var u = o[s],
-                c = u.inside,
-                g = !!u.lookbehind,
-                h = !!u.greedy,
-                f = 0,
-                d = u.alias;
-
-            if (h && !u.pattern.global) {
-              var p = u.pattern.toString().match(/[imuy]*$/)[0];
-              u.pattern = RegExp(u.pattern.source, p + "g");
-            }
-
-            u = u.pattern || u;
-
-            for (var m = 0, y = 0; m < r.length; y += (r[m].matchedStr || r[m]).length, ++m) {
-              var v = r[m];
-              if (r.length > e.length) break e;
-
-              if (!(v instanceof a)) {
-                u.lastIndex = 0;
-                var b = u.exec(v),
-                    k = 1;
-
-                if (!b && h && m != r.length - 1) {
-                  if (u.lastIndex = y, b = u.exec(e), !b) break;
-
-                  for (var w = b.index + (g ? b[1].length : 0), _ = b.index + b[0].length, A = m, S = y, P = r.length; P > A && _ > S; ++A) {
-                    S += (r[A].matchedStr || r[A]).length, w >= S && (++m, y = S);
-                  }
-
-                  if (r[m] instanceof a || r[A - 1].greedy) continue;
-                  k = A - m, v = e.slice(y, S), b.index -= y;
-                }
-
-                if (b) {
-                  g && (f = b[1].length);
-
-                  var w = b.index + f,
-                      b = b[0].slice(f),
-                      _ = w + b.length,
-                      x = v.slice(0, w),
-                      O = v.slice(_),
-                      j = [m, k];
-
-                  x && j.push(x);
-                  var N = new a(l, c ? n.tokenize(b, c) : b, d, b, h);
-                  j.push(N), O && j.push(O), Array.prototype.splice.apply(r, j);
-                }
-              }
-            }
-          }
-        }
-      }
-
-      return r;
-    },
-    hooks: {
-      all: {},
-      add: function add(e, t) {
-        var a = n.hooks.all;
-        a[e] = a[e] || [], a[e].push(t);
-      },
-      run: function run(e, t) {
-        var a = n.hooks.all[e];
-        if (a && a.length) for (var r, i = 0; r = a[i++];) {
-          r(t);
-        }
-      }
-    }
-  },
-      a = n.Token = function (e, t, n, a, r) {
-    this.type = e, this.content = t, this.alias = n, this.matchedStr = a || null, this.greedy = !!r;
-  };
-
-  if (a.stringify = function (e, t, r) {
-    if ("string" == typeof e) return e;
-    if ("Array" === n.util.type(e)) return e.map(function (n) {
-      return a.stringify(n, t, e);
-    }).join("");
-    var i = {
-      type: e.type,
-      content: a.stringify(e.content, t, r),
-      tag: "span",
-      classes: ["token", e.type],
-      attributes: {},
-      language: t,
-      parent: r
-    };
-
-    if ("comment" == i.type && (i.attributes.spellcheck = "true"), e.alias) {
-      var l = "Array" === n.util.type(e.alias) ? e.alias : [e.alias];
-      Array.prototype.push.apply(i.classes, l);
-    }
-
-    n.hooks.run("wrap", i);
-    var o = "";
-
-    for (var s in i.attributes) {
-      o += (o ? " " : "") + s + '="' + (i.attributes[s] || "") + '"';
-    }
-
-    return "<" + i.tag + ' class="' + i.classes.join(" ") + '"' + (o ? " " + o : "") + ">" + i.content + "</" + i.tag + ">";
-  }, !_self.document) return _self.addEventListener ? (_self.addEventListener("message", function (e) {
-    var t = JSON.parse(e.data),
-        a = t.language,
-        r = t.code,
-        i = t.immediateClose;
-    _self.postMessage(n.highlight(r, n.languages[a], a)), i && _self.close();
-  }, !1), _self.Prism) : _self.Prism;
-  var r = document.currentScript || [].slice.call(document.getElementsByTagName("script")).pop();
-  return r && (n.filename = r.src, document.addEventListener && !r.hasAttribute("data-manual") && ("loading" !== document.readyState ? window.requestAnimationFrame ? window.requestAnimationFrame(n.highlightAll) : window.setTimeout(n.highlightAll, 16) : document.addEventListener("DOMContentLoaded", n.highlightAll))), _self.Prism;
-}();
-
-"undefined" != typeof module && module.exports && (module.exports = Prism), "undefined" != typeof global && (global.Prism = Prism);
-Prism.languages.markup = {
-  comment: /<!--[\w\W]*?-->/,
-  prolog: /<\?[\w\W]+?\?>/,
-  doctype: /<!DOCTYPE[\w\W]+?>/i,
-  cdata: /<!\[CDATA\[[\w\W]*?]]>/i,
-  tag: {
-    pattern: /<\/?(?!\d)[^\s>\/=$<]+(?:\s+[^\s>\/=]+(?:=(?:("|')(?:\\\1|\\?(?!\1)[\w\W])*\1|[^\s'">=]+))?)*\s*\/?>/i,
-    inside: {
-      tag: {
-        pattern: /^<\/?[^\s>\/]+/i,
-        inside: {
-          punctuation: /^<\/?/,
-          namespace: /^[^\s>\/:]+:/
-        }
-      },
-      "attr-value": {
-        pattern: /=(?:('|")[\w\W]*?(\1)|[^\s>]+)/i,
-        inside: {
-          punctuation: /[=>"']/
-        }
-      },
-      punctuation: /\/?>/,
-      "attr-name": {
-        pattern: /[^\s>\/]+/,
-        inside: {
-          namespace: /^[^\s>\/:]+:/
-        }
-      }
-    }
-  },
-  entity: /&#?[\da-z]{1,8};/i
-}, Prism.hooks.add("wrap", function (a) {
-  "entity" === a.type && (a.attributes.title = a.content.replace(/&amp;/, "&"));
-}), Prism.languages.xml = Prism.languages.markup, Prism.languages.html = Prism.languages.markup, Prism.languages.mathml = Prism.languages.markup, Prism.languages.svg = Prism.languages.markup;
-Prism.languages.css = {
-  comment: /\/\*[\w\W]*?\*\//,
-  atrule: {
-    pattern: /@[\w-]+?.*?(;|(?=\s*\{))/i,
-    inside: {
-      rule: /@[\w-]+/
-    }
-  },
-  url: /url\((?:(["'])(\\(?:\r\n|[\w\W])|(?!\1)[^\\\r\n])*\1|.*?)\)/i,
-  selector: /[^\{\}\s][^\{\};]*?(?=\s*\{)/,
-  string: {
-    pattern: /("|')(\\(?:\r\n|[\w\W])|(?!\1)[^\\\r\n])*\1/,
-    greedy: !0
-  },
-  property: /(\b|\B)[\w-]+(?=\s*:)/i,
-  important: /\B!important\b/i,
-  "function": /[-a-z0-9]+(?=\()/i,
-  punctuation: /[(){};:]/
-}, Prism.languages.css.atrule.inside.rest = Prism.util.clone(Prism.languages.css), Prism.languages.markup && (Prism.languages.insertBefore("markup", "tag", {
-  style: {
-    pattern: /(<style[\w\W]*?>)[\w\W]*?(?=<\/style>)/i,
-    lookbehind: !0,
-    inside: Prism.languages.css,
-    alias: "language-css"
-  }
-}), Prism.languages.insertBefore("inside", "attr-value", {
-  "style-attr": {
-    pattern: /\s*style=("|').*?\1/i,
-    inside: {
-      "attr-name": {
-        pattern: /^\s*style/i,
-        inside: Prism.languages.markup.tag.inside
-      },
-      punctuation: /^\s*=\s*['"]|['"]\s*$/,
-      "attr-value": {
-        pattern: /.+/i,
-        inside: Prism.languages.css
-      }
-    },
-    alias: "language-css"
-  }
-}, Prism.languages.markup.tag));
-Prism.languages.clike = {
-  comment: [{
-    pattern: /(^|[^\\])\/\*[\w\W]*?\*\//,
-    lookbehind: !0
-  }, {
-    pattern: /(^|[^\\:])\/\/.*/,
-    lookbehind: !0
-  }],
-  string: {
-    pattern: /(["'])(\\(?:\r\n|[\s\S])|(?!\1)[^\\\r\n])*\1/,
-    greedy: !0
-  },
-  "class-name": {
-    pattern: /((?:\b(?:class|interface|extends|implements|trait|instanceof|new)\s+)|(?:catch\s+\())[a-z0-9_\.\\]+/i,
-    lookbehind: !0,
-    inside: {
-      punctuation: /(\.|\\)/
-    }
-  },
-  keyword: /\b(if|else|while|do|for|return|in|instanceof|function|new|try|throw|catch|finally|null|break|continue)\b/,
-  "boolean": /\b(true|false)\b/,
-  "function": /[a-z0-9_]+(?=\()/i,
-  number: /\b-?(?:0x[\da-f]+|\d*\.?\d+(?:e[+-]?\d+)?)\b/i,
-  operator: /--?|\+\+?|!=?=?|<=?|>=?|==?=?|&&?|\|\|?|\?|\*|\/|~|\^|%/,
-  punctuation: /[{}[\];(),.:]/
-};
-Prism.languages.javascript = Prism.languages.extend("clike", {
-  keyword: /\b(as|async|await|break|case|catch|class|const|continue|debugger|default|delete|do|else|enum|export|extends|finally|for|from|function|get|if|implements|import|in|instanceof|interface|let|new|null|of|package|private|protected|public|return|set|static|super|switch|this|throw|try|typeof|var|void|while|with|yield)\b/,
-  number: /\b-?(0x[\dA-Fa-f]+|0b[01]+|0o[0-7]+|\d*\.?\d+([Ee][+-]?\d+)?|NaN|Infinity)\b/,
-  "function": /[_$a-zA-Z\xA0-\uFFFF][_$a-zA-Z0-9\xA0-\uFFFF]*(?=\()/i,
-  operator: /--?|\+\+?|!=?=?|<=?|>=?|==?=?|&&?|\|\|?|\?|\*\*?|\/|~|\^|%|\.{3}/
-}), Prism.languages.insertBefore("javascript", "keyword", {
-  regex: {
-    pattern: /(^|[^\/])\/(?!\/)(\[.+?]|\\.|[^\/\\\r\n])+\/[gimyu]{0,5}(?=\s*($|[\r\n,.;})]))/,
-    lookbehind: !0,
-    greedy: !0
-  }
-}), Prism.languages.insertBefore("javascript", "string", {
-  "template-string": {
-    pattern: /`(?:\\\\|\\?[^\\])*?`/,
-    greedy: !0,
-    inside: {
-      interpolation: {
-        pattern: /\$\{[^}]+\}/,
-        inside: {
-          "interpolation-punctuation": {
-            pattern: /^\$\{|\}$/,
-            alias: "punctuation"
-          },
-          rest: Prism.languages.javascript
-        }
-      },
-      string: /[\s\S]+/
-    }
-  }
-}), Prism.languages.markup && Prism.languages.insertBefore("markup", "tag", {
-  script: {
-    pattern: /(<script[\w\W]*?>)[\w\W]*?(?=<\/script>)/i,
-    lookbehind: !0,
-    inside: Prism.languages.javascript,
-    alias: "language-javascript"
-  }
-}), Prism.languages.js = Prism.languages.javascript;
-!function () {
-  if (("undefined" == typeof self || self.Prism) && ("undefined" == typeof global || global.Prism)) {
-    var i = /\b([a-z]{3,7}:\/\/|tel:)[\w\-+%~\/.:#=?&amp;]+/,
-        n = /\b\S+@[\w.]+[a-z]{2}/,
-        e = /\[([^\]]+)]\(([^)]+)\)/,
-        t = ["comment", "url", "attr-value", "string"];
-    Prism.plugins.autolinker = {
-      processGrammar: function processGrammar(a) {
-        a && !a["url-link"] && (Prism.languages.DFS(a, function (a, r, l) {
-          t.indexOf(l) > -1 && "Array" !== Prism.util.type(r) && (r.pattern || (r = this[a] = {
-            pattern: r
-          }), r.inside = r.inside || {}, "comment" == l && (r.inside["md-link"] = e), "attr-value" == l ? Prism.languages.insertBefore("inside", "punctuation", {
-            "url-link": i
-          }, r) : r.inside["url-link"] = i, r.inside["email-link"] = n);
-        }), a["url-link"] = i, a["email-link"] = n);
-      }
-    }, Prism.hooks.add("before-highlight", function (i) {
-      Prism.plugins.autolinker.processGrammar(i.grammar);
-    }), Prism.hooks.add("wrap", function (i) {
-      if (/-link$/.test(i.type)) {
-        i.tag = "a";
-        var n = i.content;
-        if ("email-link" == i.type && 0 != n.indexOf("mailto:")) n = "mailto:" + n;else if ("md-link" == i.type) {
-          var t = i.content.match(e);
-          n = t[2], i.content = t[1];
-        }
-        i.attributes.href = n;
-      }
-    });
-  }
-}();
-!function () {
-  if ("undefined" != typeof self && self.Prism && self.document) {
-    var e = {
-      html: "HTML",
-      xml: "XML",
-      svg: "SVG",
-      mathml: "MathML",
-      css: "CSS",
-      clike: "C-like",
-      javascript: "JavaScript",
-      abap: "ABAP",
-      actionscript: "ActionScript",
-      apacheconf: "Apache Configuration",
-      apl: "APL",
-      applescript: "AppleScript",
-      asciidoc: "AsciiDoc",
-      aspnet: "ASP.NET (C#)",
-      autoit: "AutoIt",
-      autohotkey: "AutoHotkey",
-      basic: "BASIC",
-      csharp: "C#",
-      cpp: "C++",
-      coffeescript: "CoffeeScript",
-      "css-extras": "CSS Extras",
-      fsharp: "F#",
-      glsl: "GLSL",
-      graphql: "GraphQL",
-      http: "HTTP",
-      inform7: "Inform 7",
-      json: "JSON",
-      latex: "LaTeX",
-      livescript: "LiveScript",
-      lolcode: "LOLCODE",
-      matlab: "MATLAB",
-      mel: "MEL",
-      nasm: "NASM",
-      nginx: "nginx",
-      nsis: "NSIS",
-      objectivec: "Objective-C",
-      ocaml: "OCaml",
-      parigp: "PARI/GP",
-      php: "PHP",
-      "php-extras": "PHP Extras",
-      powershell: "PowerShell",
-      properties: ".properties",
-      protobuf: "Protocol Buffers",
-      jsx: "React JSX",
-      rest: "reST (reStructuredText)",
-      sas: "SAS",
-      sass: "Sass (Sass)",
-      scss: "Sass (Scss)",
-      sql: "SQL",
-      typescript: "TypeScript",
-      vhdl: "VHDL",
-      vim: "vim",
-      wiki: "Wiki markup",
-      xojo: "Xojo (REALbasic)",
-      yaml: "YAML"
-    };
-    Prism.hooks.add("before-highlight", function (s) {
-      var a = s.element.parentNode;
-
-      if (a && /pre/i.test(a.nodeName)) {
-        var t,
-            i,
-            r = a.getAttribute("data-language") || e[s.language] || s.language.substring(0, 1).toUpperCase() + s.language.substring(1),
-            p = a.previousSibling;
-        p && /\s*\bprism-show-language\b\s*/.test(p.className) && p.firstChild && /\s*\bprism-show-language-label\b\s*/.test(p.firstChild.className) ? i = p.firstChild : (t = document.createElement("div"), i = document.createElement("div"), i.className = "prism-show-language-label", t.className = "prism-show-language", t.appendChild(i), a.parentNode.insertBefore(t, a)), i.innerHTML = r;
-      }
-    });
-  }
-}();
-
-}).call(this,typeof global !== "undefined" ? global : typeof self !== "undefined" ? self : typeof window !== "undefined" ? window : {})
-},{}]},{},[4]);
+export { Remarkable, utils };
